This tag is working copy of openairinterface with abstraction especially for MU-MIMO.

+++ openair1/SIMULATION/TOOLS/multipath_channel.c
@@ -129,7 +129,7 @@ void multipath_channel(channel_desc_t *desc,
 		       double **tx_sig_im, 
 		       double **rx_sig_re,
 		       double **rx_sig_im,
-		       u32 length,
+		       u16 length,
 		       u8 keep_channel) {
  
   int i,ii,j,l;
